requirements.txt
biopython
joblib
more-itertools
numpy
pandas
pickleshare
pyrodigal
scipy
tables
xgboost
cachier
